requirements.txt
google-genai>=1.0.0
pillow>=10.0.0
streamlit>=1.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
jinja2>=3.0
tenacity>=8.0
python-dotenv>=1.0.0
